requirements.txt
pandas
numpy
scipy
requests
streamlit>=1.37.0
gspread
oauth2client
pytz
psycopg2-binary
lxml
html5lib
beautifulsoup4
twilio
scikit-learn==1.3.2
python-dotenv
nba_api
kenpompy
mechanicalsoup
bs4
joblib
pybaseball
selenium
webdriver_manager
tweepy
xgboost
fastapi
uvicorn
